General tidying up: - libconfuse needs long int's when initializing options (fixes mem_hardlimit bug) - libnids needs int's in sniffing parameters (cast long int's back to int's) - fix documentation about mem_hardlimit / mem_softlimit - make Warnings break build in sensor/ and fix all warnings

+++ sensor/sensor.h
@@ -6,10 +6,10 @@
 /* Program variables */
 typedef struct _ProgVars {
 	char *prog_name;
-	int agent_port;
-	int max_agents;
-	int mem_softlimit;
-	int mem_hardlimit;
+	long agent_port;
+	long max_agents;
+	long mem_softlimit;
+	long mem_hardlimit;
 	char *password;
 	uint8_t port_table[65536];
 } PV;
